Refactored HTTP client request and response.
