+++ libs/network/test/http/client/request_test.cpp
@@ -3,14 +3,10 @@
 // (See accompanying file LICENSE_1_0.txt or copy at
 // http://www.boost.org/LICENSE_1_0.txt)
 
-
 #include <gtest/gtest.h>
-// #include <boost/network/protocol/http/client/request.hpp>
-#include <boost/network/protocol/http/client.hpp>
+#include <boost/network/protocol/http/client/request.hpp>
 
-
-using client_request = boost::network::http::client::request;
-
+using client_request = boost::network::http::client_request;
 
 TEST(http_client_request, construct_from_uri) {
   ASSERT_NO_THROW(client_request("http://cpp-netlib.org/"));
